+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## master
 
+-   Fix witness unsupported key logger bug
+-   Upgrade tendermint/tm-db to v0.2.0
+-   Fix Fix witness connection mngmt and error handling
+-   Remove confirmation_threshold from ConsensusParams
+-   Ethereum Snapshot block
+
 ## v0.3.0
 
 -   introduce validators verification from app_state